add settlement account for supplier

+++ app/Http/Controllers/SupplierController.php
@@ -10,6 +10,9 @@ use Illuminate\Http\Request;
 class SupplierController extends Controller
 {
     public function suppliers(Request $request) {
+        $accounts = Account::where('level', 5)->get();
+        $supplierAccount = Account::where('name', 'الموردين')->first();
+        $supplierTypes = $supplierAccount ? $supplierAccount->children()->pluck('name')->toArray() : [];
         $suppliers = Supplier::query();
         
         $search = $request->input('search', null);
@@ -23,11 +26,17 @@ class SupplierController extends Controller
         $accountId = Account::where('name', 'الموردين')->first()->id;
         $supplierAccounts = Account::where('parent_id', $accountId)->get();
 
-        return view('pages.suppliers.suppliers', compact('suppliers', 'supplierAccounts'));
+        return view('pages.suppliers.suppliers', compact(
+            'suppliers', 
+            'supplierAccounts', 
+            'accounts',
+            'supplierTypes',
+        ));
     }
 
     public function supplierProfile(Supplier $supplier) {
-        return view('pages.suppliers.supplierProfile', compact('supplier'));
+        $accounts = Account::where('level', 5)->get();
+        return view('pages.suppliers.supplier_profile', compact('supplier', 'accounts'));
     }
 
     public function storeSupplier(SupplierRequest $request) {
@@ -35,7 +44,7 @@ class SupplierController extends Controller
             return redirect()->back()->with('error', 'يرجى إنشاء حساب حسابات دائنة تحت التسوية أولاً من شاشة الحسابات');
         }
 
-        $accountId = Account::where('name', 'حسابات دائنة تحت التسوية')->where('level', 4)->first()->id;
+        $accountId = Account::where('name', $request->type)->where('level', 4)->first()->id;
         $lastSupplier = Account::where('parent_id', $accountId)->latest('id')->first();
         if($lastSupplier) {
             $code = (string)((int)$lastSupplier->code + 1);
@@ -44,16 +53,32 @@ class SupplierController extends Controller
             $code = $code . '0001';
         }
         $account = Account::create([
-            'name' => $request->name . " - تحت التسوية",
+            'name' => $request->name,
             'code' => $code,
             'parent_id' => $accountId,
             'type_id' => 1,
             'level' => 5
         ]);
-        logActivity('إنشاء حساب', "تم إنشاء حساب مورد جديد باسم " . $account->name, null, $account->toArray());
+
+        $settlementAccountId = Account::where('name', 'حسابات دائنة تحت التسوية')->where('level', 4)->first()->id;
+        $lastSupplier = Account::where('parent_id', $settlementAccountId)->latest('id')->first();
+        if($lastSupplier) {
+            $code = (string)((int)$lastSupplier->code + 1);
+        } else {
+            $code = Account::where('id', $settlementAccountId)->latest('id')->first()->code;
+            $code = $code . '0001';
+        }
+        $settlement_account = Account::create([
+            'name' => $request->name . " - تحت التسوية",
+            'code' => $code,
+            'parent_id' => $settlementAccountId,
+            'type_id' => 1,
+            'level' => 5
+        ]);
 
         $validated = $request->validated();
         $validated['account_id'] = $account->id;
+        $validated['settlement_account_id'] = $settlement_account->id;
         $new = Supplier::create($validated);
         logActivity('إنشاء مورد', "تم إنشاء مورد جديد باسم " . $new->name, null, $new->toArray());
 
@@ -72,7 +97,10 @@ class SupplierController extends Controller
     }
 
     public function deleteSupplier(Supplier $supplier) {
-        
+        if($supplier->expenseInvoices()->exists() || $supplier->shippingPolicies()->exists() || $supplier->transportOrders()->exists()) {
+            return redirect()->back()->with('error', 'لا يمكن حذف هذا المورد لأنه مرتبط بسجلات أخرى في النظام');
+        }
+
         $old = $supplier->toArray();
         $supplier->delete();
         logActivity('حذف مورد', "تم حذف مورد باسم " . $supplier->name, $old, null);
